256ai-projects/bazi/iOS/AstrologyApp/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
openai>=1.12.0
lunar_python>=1.4.8
pydantic[email]>=2.0.0
ollama>=0.4.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.27.0
PyJWT>=2.8.0
google-auth>=2.27.0

# Rate Limiting
slowapi>=0.1.9
